+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Automatically refresh torrent details screen [#101](https://github.com/anlar/tewi/issues/101)
 - Edit torrent name and location [#103](https://github.com/anlar/tewi/issues/103)
 - Display search hits in status panel [#104](https://github.com/anlar/tewi/issues/104)
+- Display qBittorrent-specific statistics (waste, connected peers, cache, performance) [#106](https://github.com/anlar/tewi/issues/106)
 
 ### Changed